requirements.txt
# OCR Models
pytesseract>=0.3.10
easyocr>=1.7.0
paddlepaddle==2.6.2
paddleocr==2.8.1
python-doctr[torch]>=0.8.0
transformers>=4.36.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Evaluation
jiwer>=3.0.0

# Visualization & Data
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0

# API calls (Mistral)
requests>=2.31.0

# Dataset
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
tiktoken
sentencepiece
